powerpc/perf: Fix incorrect return type for perf_misc_flags
The return value of perf_misc_flags is assigned to the `misc`
field of the perf_event_header structure, and its type should
be `unsigned short`.

Fixes: a1110654 ("powerpc/perf: e500 support")
Fixes: 98fb180 ("perf_counter: powerpc: Make powerpc
perf_counter code safe for 32-bit kernels")

+++ arch/powerpc/include/asm/perf_event_server.h
@@ -102,7 +102,7 @@ struct power_pmu {
 int __init register_power_pmu(struct power_pmu *pmu);
 
 struct pt_regs;
-extern unsigned long perf_misc_flags(struct pt_regs *regs);
+extern unsigned short perf_misc_flags(struct pt_regs *regs);
 extern unsigned long perf_instruction_pointer(struct pt_regs *regs);
 extern unsigned long int read_bhrb(int n);
